+++ AngularApp/projects/app-service-diagnostics/src/app/shared/components/tools/network-checks/apim-flows/DnsFlow.ts
@@ -1,19 +1,11 @@
-import { InfoStepView, InfoType, StatusStyles } from 'diagnostic-data';
+import { InfoStepView, InfoType } from 'diagnostic-data';
 import { NetworkCheckFlow } from "../network-check-flow";
+import { APIM_API_VERSION } from './data/constants';
 import { dnsMismatchCheck } from './checks/dnsMismatchCheck';
 import { networkStatusCheck } from './checks/networkStatusCheck';
 import { nsgRuleCheck } from './checks/nsgCheck';
 import { routeTableCheck } from './checks/routeTableCheck';
 import { ApiManagementServiceResourceContract } from './contracts/APIMService';
-
-export const APIM_API_VERSION = "2021-12-01-preview";
-export const NETWORK_API_VERSION = "2021-08-01";
-
-export let statusIconMarkdown = {
-    0: `<i class="${StatusStyles.HealthyIcon}" style="width: 17px; height: 17px;"></i> Success`,
-    1: `<i class="${StatusStyles.WarningIcon}" style="width: 17px; height: 17px;"></i> Warning`,
-    2: `<i class="${StatusStyles.CriticalIcon}" style="width: 17px; height: 17px;"></i> Error`, // fail
-};
 
 export const DnsFlow: NetworkCheckFlow = {
     title: "Network Connectivity Check",
